# BFD-2247: Update some Patient SLO thresholds, and improve error SLOs and implementation